rename auth models.py to table.py
